layers_4-6/monitoring/requirements.txt
psutil
pyserial
psycopg2-binary
argparse
